+++ app/app.go
@@ -11,9 +11,13 @@ import (
 	"github.com/gorilla/mux"
 	"github.com/rakyll/statik/fs"
 	feemarketkeeper "github.com/skip-mev/feemarket/x/feemarket/keeper"
+	"github.com/spf13/cast"
+	"github.com/spf13/viper"
 
 	abci "github.com/cometbft/cometbft/abci/types"
+	tmcfg "github.com/cometbft/cometbft/config"
 	tmjson "github.com/cometbft/cometbft/libs/json"
+	"github.com/cometbft/cometbft/privval"
 	tmproto "github.com/cometbft/cometbft/proto/tendermint/types"
 
 	dbm "github.com/cosmos/cosmos-db"
@@ -37,6 +41,7 @@ import (
 
 	"github.com/cosmos/cosmos-sdk/baseapp"
 	"github.com/cosmos/cosmos-sdk/client"
+	"github.com/cosmos/cosmos-sdk/client/flags"
 	"github.com/cosmos/cosmos-sdk/client/grpc/cmtservice"
 	nodeservice "github.com/cosmos/cosmos-sdk/client/grpc/node"
 	"github.com/cosmos/cosmos-sdk/codec"
@@ -61,6 +66,7 @@ import (
 	authtx "github.com/cosmos/cosmos-sdk/x/auth/tx"
 	txmodule "github.com/cosmos/cosmos-sdk/x/auth/tx/config"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
+	genutiltypes "github.com/cosmos/cosmos-sdk/x/genutil/types"
 	govkeeper "github.com/cosmos/cosmos-sdk/x/gov/keeper"
 	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
 
@@ -72,6 +78,8 @@ import (
 	"github.com/cosmos/gaia/v25/app/keepers"
 	"github.com/cosmos/gaia/v25/app/upgrades"
 	v25 "github.com/cosmos/gaia/v25/app/upgrades/v25_0_0"
+	gaiatelemetry "github.com/cosmos/gaia/v25/telemetry"
+	"github.com/cosmos/gaia/v25/x/telemetry"
 )
 
 var (
@@ -106,6 +114,8 @@ type GaiaApp struct { //nolint: revive
 	// simulation manager
 	sm           *module.SimulationManager
 	configurator module.Configurator
+
+	otelClient *gaiatelemetry.OtelClient
 }
 
 func init() {
@@ -170,6 +180,23 @@ func NewGaiaApp(
 		interfaceRegistry: interfaceRegistry,
 	}
 
+	vi, err := getValidatorInfo(homePath, appOpts)
+	if err != nil {
+		logger.Debug("failed to get validator info: unable to determine if this node is a validator", "err", err)
+	} else {
+		logger.Debug("successfully determined if this node is a validator", "moniker", vi.Moniker)
+	}
+
+	otelConfig := gaiatelemetry.OtelConfig{
+		Disable:                 cast.ToBool(appOpts.Get("opentelemetry.disable")),
+		CollectorEndpoint:       cast.ToString(appOpts.Get("opentelemetry.collector-endpoint")),
+		CollectorMetricsURLPath: cast.ToString(appOpts.Get("opentelemetry.collector-metrics-url-path")),
+		User:                    cast.ToString(appOpts.Get("opentelemetry.user")),
+		Token:                   cast.ToString(appOpts.Get("opentelemetry.token")),
+		PushInterval:            cast.ToDuration(appOpts.Get("opentelemetry.push-interval")),
+	}
+	app.otelClient = gaiatelemetry.NewOtelClient(otelConfig, vi)
+
 	moduleAccountAddresses := app.ModuleAccountAddrs()
 
 	// Setup keepers
@@ -221,6 +248,7 @@ func NewGaiaApp(
 	app.mm.SetOrderPreBlockers(
 		upgradetypes.ModuleName,
 		authtypes.ModuleName,
+		telemetry.ModuleName,
 	)
 	// During begin block slashing happens after distr.BeginBlocker so that
 	// there is nothing left over in the validator fee pool, so as to keep the
@@ -356,6 +384,15 @@ func NewGaiaApp(
 		if err := app.WasmClientKeeper.InitializePinnedCodes(ctx); err != nil {
 			panic(fmt.Sprintf("wasmlckeeper failed initialize pinned codes %s", err))
 		}
+	}
+
+	if otelConfig.CollectorEndpoint != "" && !otelConfig.Disable {
+		logger.Debug("creating gaia app with open telemetry")
+		if err := app.otelClient.StartExporter(logger); err != nil {
+			panic(err)
+		}
+	} else {
+		logger.Debug("creating gaia app without open telemetry")
 	}
 
 	return app
@@ -567,6 +604,48 @@ func (app *GaiaApp) AutoCliOpts() autocli.AppOptions {
 	}
 }
 
+func getValidatorInfo(homePath string, appOpts servertypes.AppOptions) (gaiatelemetry.ValidatorInfo, error) {
+	cfg := &tmcfg.Config{
+		BaseConfig:      tmcfg.BaseConfig{},
+		RPC:             &tmcfg.RPCConfig{},
+		P2P:             &tmcfg.P2PConfig{},
+		Mempool:         &tmcfg.MempoolConfig{},
+		StateSync:       &tmcfg.StateSyncConfig{},
+		BlockSync:       &tmcfg.BlockSyncConfig{},
+		Consensus:       &tmcfg.ConsensusConfig{},
+		Storage:         &tmcfg.StorageConfig{},
+		TxIndex:         &tmcfg.TxIndexConfig{},
+		Instrumentation: &tmcfg.InstrumentationConfig{},
+	}
+	cfg.SetRoot(homePath)
+
+	configPath := filepath.Join(homePath, "config", "config.toml")
+	if _, err := os.Stat(configPath); err == nil {
+		viper := viper.New()
+		viper.SetConfigType("toml")
+		viper.SetConfigFile(configPath)
+		if err := viper.ReadInConfig(); err == nil {
+			if err := viper.Unmarshal(cfg); err != nil {
+				return gaiatelemetry.ValidatorInfo{}, fmt.Errorf("failed to unmarshal config file: %w", err)
+			}
+		}
+	} else {
+		return gaiatelemetry.ValidatorInfo{}, fmt.Errorf("unable to stat config file at %s", configPath)
+	}
+
+	chainID := cast.ToString(appOpts.Get(flags.FlagChainID))
+	if chainID == "" {
+		genDocFile := filepath.Join(homePath, "config", "genesis.json")
+		appGenesis, err := genutiltypes.AppGenesisFromFile(genDocFile)
+		if err == nil {
+			chainID = appGenesis.ChainID
+		}
+	}
+
+	vi, err := validatorInfoFromCometConfig(cfg, chainID)
+	return vi, err
+}
+
 // TestingApp functions
 
 // GetBaseApp implements the TestingApp interface.
@@ -632,4 +711,42 @@ func minTxFeesChecker(ctx sdk.Context, tx sdk.Tx, feemarketKp feemarketkeeper.Ke
 	}
 
 	return feeTx.GetFee(), 0, nil
+}
+
+var ErrNotValidator = fmt.Errorf("not validator")
+
+func validatorInfoFromCometConfig(cfg *tmcfg.Config, chainID string) (gaiatelemetry.ValidatorInfo, error) {
+	vi := gaiatelemetry.ValidatorInfo{
+		ChainID: chainID,
+	}
+	if cfg.PrivValidatorListenAddr != "" {
+		listenAddr := cfg.PrivValidatorListenAddr
+		pve, err := privval.NewSignerListener(listenAddr, nil)
+		if err != nil {
+			return vi, fmt.Errorf("failed to start private validator: %w", err)
+		}
+
+		pvsc, err := privval.NewSignerClient(pve, chainID)
+		if err != nil {
+			return vi, fmt.Errorf("failed to start private validator: %w", err)
+		}
+
+		pk, err := pvsc.GetPubKey()
+		if err != nil {
+			return vi, fmt.Errorf("cannot get pubkey from remote signer: %w", err)
+		}
+		vi.Moniker = cfg.Moniker
+		vi.Address = pk.Address()
+		return vi, nil
+	} else if cfg.PrivValidatorKey != "" {
+		vi.Moniker = cfg.Moniker
+		_, err := os.Stat(cfg.PrivValidatorKeyFile())
+		if err != nil {
+			return vi, ErrNotValidator
+		}
+		pv := privval.LoadFilePV(cfg.PrivValidatorKeyFile(), cfg.PrivValidatorStateFile())
+		vi.Address = pv.GetAddress()
+		return vi, nil
+	}
+	return vi, ErrNotValidator
 }