+++ app/sim_test.go
@@ -55,6 +55,7 @@ func TestAppStateDeterminism(t *testing.T) {
 	config := sim.NewConfigFromFlags()
 	config.InitialBlockHeight = 1
 	config.ExportParamsPath = ""
+	// TODO: update deprecated code.
 	config.OnOperation = false   //nolint:staticcheck
 	config.AllInvariants = false //nolint:staticcheck
 	config.GenesisTime = time.Now().UTC().Unix()